Significant refactoring to put most of the generic stuff in its own modules, as well as to clean up some repetitious code.
