+++ .gitignore
@@ -1,5 +1,6 @@
 **/.DS_Store
+.vscode
 output
 coverage
 resources
-*log*
+*.log
