+++ lib/generic/modules/example_module_alt.json
